Упорядочивает константы
