+++ src/const.js
@@ -1,19 +1,7 @@
 export const CITIES = [`Paris`, `Cologne`, `Brussels`, `Amsterdam`, `Hamburg`, `Dusseldorf`];
-export const CURRENT_CITY = `Amsterdam`;
-export const OFFER_TYPES = [`apartment`, `room`, `house`, `hotel`];
 
-export const CardType = {
+export const OfferType = {
   MAIN: `cities`,
   NEIGHBOURHOOD: `near-places`,
   FAVORITES: `favorites`
-};
-
-export const MAX_RATING = 5;
-
-export const RatingValueToTitle = {
-  "1": `terribly`,
-  "2": `badly`,
-  "3": `not bad`,
-  "4": `good`,
-  "5": `perfect`
 };
