+++ src/util/mock.js
@@ -1,6 +1,7 @@
 const DESCRIPTION_MOCK = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`;
 
 const MAX_SENTENCES_COUNT = 4;
+const MAX_RATING = 5;
 
 const getRandomInteger = (a = 0, b = 1) => {
   const lower = Math.ceil(Math.min(a, b));
@@ -26,4 +27,8 @@ export const generateDescription = () => {
   }).join(` `);
 
   return description;
+};
+
+export const getRatingInPercent = (ratingValue) => {
+  return Math.floor(100 * ratingValue / MAX_RATING);
 };
